gdiplus: Fix incorrect pointer cast writing pens to emf+.
CPY_34
Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=60241

+++ dlls/gdiplus/metafile.c
@@ -5051,7 +5051,7 @@ static GpStatus METAFILE_AddPenObject(GpMetafile *metafile, GpPen *pen, DWORD *i
     }
     if (data_flags & PenDataNonCenter)
     {
-        *(REAL*)(pen_data->OptionalData + i) = pen->align;
+        *(DWORD*)(pen_data->OptionalData + i) = pen->align;
         i += sizeof(DWORD);
     }
     if (data_flags & PenDataCustomStartCap)
